v5.x-scala-js: move SigningSpecification to interpreter
